epsilon is double

+++ kmeans.c
@@ -13,7 +13,7 @@ typedef struct
 } Cluster;
 
 /*Prototypes*/
-static PyObject *kmeans(int k, int max_iter_py, int epsilon_py, int dim_py, int N_py, PyObject *centroids_py, PyObject *vectors_py);
+static PyObject *kmeans(int k, int max_iter_py, double epsilon_py, int dim_py, int N_py, PyObject *centroids_py, PyObject *vectors_py);
 void calcCluster(double* vector, Cluster* clusters, int k, int dim);
 double calcDistance(double* vector1, double* vector2, int dim);
 int updateCentroids(Cluster* clusters, int k, int dim, double epsilon);
@@ -27,11 +27,11 @@ int main(int argc, char *argv[])
 }
 
 
-static PyObject *kmeans(int k, int max_iter_py, int epsilon_py, int dim_py, int N_py, PyObject *centroids_py, PyObject *vectors_py)
+static PyObject *kmeans(int k, int max_iter_py, double epsilon_py, int dim_py, int N_py, PyObject *centroids_py, PyObject *vectors_py)
 {
     int N = N_py;
     int max_iter = max_iter_py;
-    int epsilon = epsilon_py;
+    double epsilon = epsilon_py;
     int dim = dim_py;
     Cluster *clusters;
     double *curr_vector;
